02/assignment/blog_project/requirements.txt
awsebcli==3.20.10
blessed==1.20.0
botocore==1.31.85
cement==2.8.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
colorama==0.4.3
cryptography==41.0.5
defusedxml==0.7.1
Django==2.2
django-allauth==0.44.0
django-braces==1.14.0
django-ebhealthcheck==2.0.2
django-widget-tweaks==1.4.8
idna==3.4
jmespath==1.0.1
oauthlib==3.2.2
pathspec==0.10.1
Pillow==10.1.0
pycparser==2.21
PyJWT==2.8.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
semantic-version==2.8.5
six==1.16.0
sqlparse==0.4.4
termcolor==1.1.0
urllib3==1.26.18
wcwidth==0.1.9
